backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
email-validator>=2.0.0
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
livekit-api>=0.8.2
sqlalchemy>=2.0.32
psycopg[binary]>=3.2.1
